# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.0.0",
    "title": "Aplikacja Demonostracyjna",
    "description": "Aplikacja zaweira podatnosci: Prototype Pollution, NoSql Injection, Web Cache Deception",
    "contact": {
      "name": "API Support",
      "email": "hubertczarnowski"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/profile/profile": {
      "get": {
        "tags": [
          "Web Cache Deception"
        ],
        "summary": "Send Note parameters",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/profile/profile/{filename}": {
      "get": {
        "tags": [
          "Web Cache Deception"
        ],
        "summary": "Get File",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}